# Disable default paging in list watches