@dims 2017-09-04
Looks like a legit verify failure - `FAILED   hack/make-rules/../../hack/verify-bazel.sh	20s`