@smarterclayton 2017-09-06
Metrics from the two runs:

Before

```
I0905 22:34:40.939]   /go/src/k8s.io/kubernetes/_output/dockerized/go/src/k8s.io/kubernetes/test/e2e/scalability/load.go:103
I0905 22:34:41.038] Sep  5 22:34:41.037: INFO: Top latency metric: {Resource:pods Subresource: Verb:LIST Latency:{Perc50:9.261ms Perc90:11.478ms Perc99:48.501ms Perc100:0s} Count:9840}
I0905 22:34:41.038] Sep  5 22:34:41.037: INFO: Top latency metric: {Resource:services Subresource: Verb:DELETE Latency:{Perc50:12.146ms Perc90:15.141ms Perc99:35.373ms Perc100:0s} Count:821}
I0905 22:34:41.038] Sep  5 22:34:41.037: INFO: Top latency metric: {Resource:nodes Subresource: Verb:LIST Latency:{Perc50:23.318ms Perc90:25.504ms Perc99:27.916ms Perc100:0s} Count:46}
I0905 22:34:41.038] Sep  5 22:34:41.037: INFO: Top latency metric: {Resource:services Subresource: Verb:POST Latency:{Perc50:3.721ms Perc90:4.891ms Perc99:18.715ms Perc100:0s} Count:821}
I0905 22:34:41.039] Sep  5 22:34:41.037: INFO: Top latency metric: {Resource:services Subresource: Verb:LIST Latency:{Perc50:15.635ms Perc90:16.778ms Perc99:16.778ms Perc100:0s} Count:10}
I0905 22:34:41.039] Sep  5 22:34:41.037: INFO: Printing summary: APIResponsiveness
```

After

```
I0906 04:26:59.107]   /go/src/k8s.io/kubernetes/_output/dockerized/go/src/k8s.io/kubernetes/test/e2e/scalability/load.go:103
I0906 04:26:59.211] Sep  6 04:26:59.211: INFO: Top latency metric: {Resource:nodes Subresource: Verb:LIST Latency:{Perc50:24.351ms Perc90:28.475ms Perc99:56.18ms Perc100:0s} Count:48}
I0906 04:26:59.212] Sep  6 04:26:59.211: INFO: Top latency metric: {Resource:pods Subresource: Verb:LIST Latency:{Perc50:10.018ms Perc90:12.068ms Perc99:49.82ms Perc100:0s} Count:9840}
I0906 04:26:59.212] Sep  6 04:26:59.211: INFO: Top latency metric: {Resource:services Subresource: Verb:DELETE Latency:{Perc50:12.294ms Perc90:15.457ms Perc99:37.681ms Perc100:0s} Count:821}
I0906 04:26:59.212] Sep  6 04:26:59.211: INFO: Top latency metric: {Resource:replicationcontrollers Subresource: Verb:LIST Latency:{Perc50:22.8ms Perc90:22.8ms Perc99:22.8ms Perc100:0s} Count:2}
I0906 04:26:59.212] Sep  6 04:26:59.211: INFO: Top latency metric: {Resource:services Subresource: Verb:LIST Latency:{Perc50:15.143ms Perc90:20.35ms Perc99:20.35ms Perc100:0s} Count:10}
I0906 04:26:59.213] Sep  6 04:26:59.211: INFO: Printing summary: APIResponsiveness
```