@wojtek-t 2017-09-06
Slightly higher, but the difference isn't large. I think at that level of latencies (small tens of ms), that may be expected. WDYT?