@smarterclayton 2017-09-06
I was expecting to see a drop (since at the apiserver a paged list should be proportionally faster) on at least one of the high N resource types.  Pods would be most likely.  However, pods are likely dominated by node list watches, not by master list watches, so pod tail latency should go down.