@smarterclayton 2017-09-06
What is the max resource collection size on the cluster?  I.e. how big do pods get at any one time?  1k? 9k?