@smarterclayton 2017-09-07
Ok, so with the second run a number of mutation operations had lower tail latencies.  This would be expected when there are conflicting reads and writes (in etcd3 at the moment there are a few range locks that large range reads can take that block writes).  However, paging wasn't happening frequently enough in the run to tell one way or another, because in practice this test doesn't ever reLIST - all caches start empty and are fed by watches.  So I'm going to say we need a better test scenario for this before we can say one way or another.

Going to drop the last commit and get everything green, then reapply label to disable paging on the client side (as the PR originally mentions).